fix(providers): correct video FPS type and update thinking logic
- Change video_fps serialization to float across Bailian, OpenAI, and
  Volcengine providers to support fractional FPS values.
- Update Volcengine thinking logic to explicitly disable thinking when
  no effort is specified, as some models reject "auto".
- Introduce SeedanceTaskError to propagate structured task failure info.

+++ backend/app/providers/bailian/base.py
@@ -325,7 +325,7 @@ class BailianProvider(OpenAIProvider):
 
         if block.type == ContentType.VIDEO_URL and block.video_fps is not None:
             item = {"type": "video_url", "video_url": {"url": block.url}}
-            item["fps"] = int(block.video_fps)
+            item["fps"] = float(block.video_fps)
             return item
         return super()._content_block_to_openai(block)
